requirements.txt
# Core Python libraries
requests
beautifulsoup4
sentence-transformers
numpy
crawl4ai
langchain
langchain-community
keybert
nltk
rank_bm25
pymongo
openai
pypdf
tiktoken
langchain-community
langchain-mongodb
unstructured[md]
sentence-transformers
langchain-huggingface
trafilatura
keybert
furo
sphinx
sphinx-rtd-theme
myst-parser
python-dotenv
pymongo